# api-docs (1).json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/profesores/actualizar/{idPersona}":{"put":{"tags":["profesor-controller"],"operationId":"actualizarProfesor","parameters":[{"name":"idPersona","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"nombre","in":"query","required":true,"schema":{"type":"string"}},{"name":"apellido","in":"query","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","required":true,"schema":{"type":"string"}},{"name":"especialidad","in":"query","required":true,"schema":{"type":"string"}},{"name":"fechaContratacion","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/personas/actualizar/{idPersona}":{"put":{"tags":["persona-controller"],"operationId":"actualizarPersona","parameters":[{"name":"idPersona","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"nombre","in":"query","required":true,"schema":{"type":"string"}},{"name":"apellido","in":"query","required":true,"schema":{"type":"string"}},{"name":"fechaNacimiento","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"email","in":"query","required":true,"schema":{"type":"string"}},{"name":"telefono","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/estudiantes/actualizar/{id}":{"put":{"tags":["estudiante-controller"],"operationId":"actualizarEstudiante","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"nombre","in":"query","required":true,"schema":{"type":"string"}},{"name":"apellido","in":"query","required":true,"schema":{"type":"string"}},{"name":"fechaNacimiento","in":"query","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","required":true,"schema":{"type":"string"}},{"name":"telefono","in":"query","required":true,"schema":{"type":"string"}},{"name":"numeroMatricula","in":"query","required":true,"schema":{"type":"string"}},{"name":"grado","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/profesores/registrar":{"post":{"tags":["profesor-controller"],"operationId":"registrarProfesor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Profesor"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profesor"}}}}}}},"/api/personas/registrar":{"post":{"tags":["persona-controller"],"operationId":"registrarPersona","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Persona"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Persona"}}}}}}},"/api/estudiantes/registrar":{"post":{"tags":["estudiante-controller"],"operationId":"registrarEstudiante","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Estudiante"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/profesores":{"get":{"tags":["profesor-controller"],"operationId":"listarProfesores","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Profesor"}}}}}}}},"/api/profesores/buscar/id/{idPersona}":{"get":{"tags":["profesor-controller"],"operationId":"buscarPorId","parameters":[{"name":"idPersona","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/profesores/buscar/especialidad/{especialidad}":{"get":{"tags":["profesor-controller"],"operationId":"buscarPorEspecialidad","parameters":[{"name":"especialidad","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/personas":{"get":{"tags":["persona-controller"],"operationId":"listarPersonas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Persona"}}}}}}}},"/api/personas/buscar/id/{idPersona}":{"get":{"tags":["persona-controller"],"operationId":"buscarPorId_1","parameters":[{"name":"idPersona","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/estudiantes":{"get":{"tags":["estudiante-controller"],"operationId":"listarEstudiantes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Estudiante"}}}}}}}},"/api/estudiantes/buscar/matricula/{numeroMatricula}":{"get":{"tags":["estudiante-controller"],"operationId":"buscarPorNumeroMatricula","parameters":[{"name":"numeroMatricula","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/estudiantes/buscar/id/{id}":{"get":{"tags":["estudiante-controller"],"operationId":"buscarPorId_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/profesores/eliminar/{idPersona}":{"delete":{"tags":["profesor-controller"],"operationId":"eliminarProfesor","parameters":[{"name":"idPersona","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/personas/{idPersona}":{"delete":{"tags":["persona-controller"],"operationId":"eliminarPersona","parameters":[{"name":"idPersona","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/estudiantes/eliminar/{id}":{"delete":{"tags":["estudiante-controller"],"operationId":"eliminarEstudiante","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"Profesor":{"type":"object","properties":{"idPersona":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"apellido":{"type":"string"},"fechaNacimiento":{"type":"string","format":"date"},"email":{"type":"string"},"telefono":{"type":"string"},"especialidad":{"type":"string"},"fechaContratacion":{"type":"string","format":"date"}}},"Persona":{"type":"object","properties":{"idPersona":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"apellido":{"type":"string"},"fechaNacimiento":{"type":"string","format":"date"},"email":{"type":"string"},"telefono":{"type":"string"}}},"Estudiante":{"type":"object","properties":{"idPersona":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"apellido":{"type":"string"},"fechaNacimiento":{"type":"string","format":"date"},"email":{"type":"string"},"telefono":{"type":"string"},"numeroMatricula":{"type":"string"},"grado":{"type":"string"}}}}}}